Add text resource reader utils program

+++ src/main/java/com/openglesbook/common/ESTransform.java
@@ -45,8 +45,8 @@ public class ESTransform {
         float sinAngle, cosAngle;
         float mag = (float) Math.sqrt((double) (x * x + y * y + z * z));
 
-        sinAngle = (float) Math.sin((double) (angle * Math.PI / 180.0));
-        cosAngle = (float) Math.cos((double) (angle * Math.PI / 180.0));
+        sinAngle = (float) Math.sin(angle * Math.PI / 180.0);
+        cosAngle = (float) Math.cos(angle * Math.PI / 180.0);
         if (mag > 0.0f) {
             float xx, yy, zz, xy, yz, zx, xs, ys, zs;
             float oneMinusCos;
